requirements.txt
python-dotenv
aiohttp
aiogram
sqlalchemy